% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_export.R
\name{chr_to_list}
\alias{chr_to_list}
\title{Separated character to list}
\usage{
chr_to_list(chr, atom_sep = ";", list_sep = "|")
}
\arguments{
\item{chr}{string to convert.}

\item{atom_sep}{separator for the elements of atomic vectors.}

\item{list_sep}{separator for the elements of lists.}
}
\value{
A list (\code{NULL} if the string is empty or \code{NA}).
}
\description{
Convert a string into list, given separators.
}
\details{
If all elements of the returned list are atomic,
the list is simplified.
}
\examples{
chr_to_list("1;2;3|a;b")
chr_to_list("1;2;3")
chr_to_list("1|2|3")
chr_to_list(c("1;3", "2|4"))
chr_to_list(c("1;3", "2;4"))
}
