% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_export.R
\name{chr_to_listcol}
\alias{chr_to_listcol}
\title{Separated character to list-columns (for a data frame)}
\usage{
chr_to_listcol(df, vars, atom_sep = ";", list_sep = "|")
}
\arguments{
\item{df}{a data frame containing columns to transform into
list-columns.}

\item{vars}{columns to transform (vector of names or indexes)}

\item{atom_sep}{separator for the elements of atomic vectors.}

\item{list_sep}{separator for the elements of lists.}
}
\value{
A data frame. Columns specified in \code{vars} are
  transformed into list-columns, other are unchanged.
}
\description{
Convert the specified columns of a data frame into list-columns,
given separators.
}
\examples{
df <-
  data.frame(
    x = c("1;2|3", "a|b"),
    y = c("1;3", "2|4"),
    z = c("1;3", "2;4")
  )
chr_to_listcol(df, "x")
chr_to_listcol(df, 2:3)
}
