% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_export.R
\name{list_to_chr}
\alias{list_to_chr}
\title{Simple list to character}
\usage{
list_to_chr(x, atom_sep = ";", list_sep = "|", simplify_threshold = Inf,
  simplify_sep = "->")
}
\arguments{
\item{x}{a list, whose lower level elements are atomic vectors.}

\item{atom_sep}{separator for the elements of atomic vectors.}

\item{list_sep}{separator for the elements of lists.}

\item{simplify_threshold}{non-negative number of atomic elements from
which the character string is simplified (only first and last elements
separated by a separator, "->" by default).}

\item{simplify_sep}{a separator (one or several characters)
if the string has to be simplified.}
}
\value{
A string, the text representation of x.
}
\description{
Returns a text representation of a list, whose lower level
elements are atomic vectors.
}
\details{
Even if the list has several hierarchical levels, the
returned value is always a string. To preserve a complex
hierarchy, use lapply or purr::map_chr at a given
hierarchical level.
}
\examples{
list_to_chr(list(letters[1:2], 1:3, 4))
list_to_chr(list(letters, LETTERS), simplify_threshold = 10)
}
