% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_export.R
\name{listcol_to_chr}
\alias{listcol_to_chr}
\title{Simple list to character (for a data frame)}
\usage{
listcol_to_chr(df, atom_sep = ";", list_sep = "|",
  simplify_threshold = Inf, simplify_symbol = "->")
}
\arguments{
\item{df}{un data.frame contenant des colonnes-listes.}

\item{atom_sep}{séparateur des éléments atomiques.}

\item{list_sep}{séparateur des éléments de la liste.}

\item{simplify_threshold}{nombre d'éléments atomiques à partir duquel
la chaîne de caractères est simplifiée (premier et dernier éléments
séparés par un symbole, "->" par défaut).}

\item{simplify_symbol}{chaîne de caractère, si la chaîne est simplifiée.}
}
\value{
Un data.frame, les colonnes simples (atomiques) restent
  inchangées, les colonnes-listes (recursives) sont transformées
  en vecteurs de chaînes de caractères.
}
\description{
Convertit toutes les colonnes-listes d'un data.frame en vecteurs
de chaînes de caractères.
}
\examples{
df <- data.frame(x = 1:2)
df$y <- list(letters[1:3], 1:2)
df$z <- list(list(3:4, 5:7), "d")
listcol_to_chr(df)
}
