% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_export.R
\name{chr_to_list}
\alias{chr_to_list}
\title{Separated character to list}
\usage{
chr_to_list(chr, atom_sep = ";", list_sep = "|")
}
\arguments{
\item{chr}{chaîne de caractères à convertir.}

\item{atom_sep}{séparateur des éléments atomiques.}

\item{list_sep}{séparateur des éléments de la liste.}
}
\value{
Une liste (NULL si la chaîne est vide ou NA).
}
\description{
Convertit une chaîne de caractère en liste,
selon des séparateurs définis.
}
\details{
Si tous les éléments de la liste en sortie sont
atomiques, la liste est simplifiée
}
\examples{
chr_to_list("1;2;3|a;b")
chr_to_list("1;2;3")
chr_to_list("1|2|3")
chr_to_list(c("1;3", "2|4"))
chr_to_list(c("1;3", "2;4"))
}
