% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_export.R
\name{chr_to_listcol}
\alias{chr_to_listcol}
\title{Separated character to list (for a data frame)}
\usage{
chr_to_listcol(df, vars, atom_sep = ";", list_sep = "|")
}
\arguments{
\item{df}{un data.frame contenant des colonnes à transformer en
colonnes-listes.}

\item{vars}{colonnes à transformer (noms ou index)}

\item{atom_sep}{séparateur des éléments atomiques.}

\item{list_sep}{séparateur des éléments de la liste.}
}
\description{
Convertit une chaîne de caractère en liste,
  selon des séparateurs définis.
}
\examples{
df <-
  data.frame(
    x = c("1;2|3", "a|b"),
    y = c("1;3", "2|4"),
    z = c("1;3", "2;4")
  )
chr_to_listcol(df, "x")
chr_to_listcol(df, 2:3)
}
