% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_export.R
\name{list_to_chr}
\alias{list_to_chr}
\title{Simple list to character}
\usage{
list_to_chr(x, atom_sep = ";", list_sep = "|", simplify_threshold = Inf,
  simplify_symbol = "->")
}
\arguments{
\item{x}{une liste dont les éléments de plus bas niveau sont des
vecteurs atomiques.}

\item{atom_sep}{séparateur des éléments atomiques.}

\item{list_sep}{séparateur des éléments de la liste.}

\item{simplify_threshold}{nombre d'éléments atomiques à partir duquel
la chaîne de caractères est simplifiée (premier et dernier éléments
séparés par un symbole, "->" par défaut).}

\item{simplify_symbol}{chaîne de caractère, si la chaîne est simplifiée.}
}
\value{
Une chaîne de caractères, la représentation texte de la liste.
}
\description{
Donne une représentation textuelle d'une liste dont les éléments
de plus bas niveau sont des vecteurs atomiques.
}
\details{
La valeur retournée est toujours une chaine, même si la liste
a plusieurs niveaux de hiérarchie. Pour conserver la hiérarchie,
appeler lapply ou purr::map_chr à un niveau hiérarchique donné.
}
\examples{
list_to_chr(list(letters[1:2], 1:3, 4))
list_to_chr(list(letters, LETTERS), simplify_threshold = 10)
}
