% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_chr.R
\name{listcol_to_chr}
\alias{listcol_to_chr}
\title{List-columns to character (for a data frame)}
\usage{
listcol_to_chr(
  df,
  atom_sep = ";",
  list_sep = "|",
  simplify_threshold = Inf,
  simplify_sep = "->"
)
}
\arguments{
\item{df}{a data frame containing list-columns.}

\item{atom_sep}{separator for the elements of atomic vectors.}

\item{list_sep}{separator for the elements of lists.}

\item{simplify_threshold}{non-negative number of atomic elements from
which the character string is simplified (only first and last elements
separated by a separator, "->" by default).}

\item{simplify_sep}{a separator (one or several characters)
if the string has to be simplified.}
}
\value{
A data frame. Simple columns (atomic) remain the same,
  list-columns (recursive) are transformed into character vectors.
}
\description{
Convert all list-columns of a data frame into character vectors.
}
\examples{
df <- data.frame(x = 1:2)
df$y <- list(letters[1:3], 1:2)
df$z <- list(list(3:4, 5:7), "d")
listcol_to_chr(df)
}
